fix(email): 修正邮箱前缀生成评审问题
- 抽出共享邮箱 local-part 生成 helper，避免 CloudMail 和 Cloudflare Temp Email 名单逻辑重复

- 将默认前缀时间精度提升到毫秒，降低高并发同秒生成时的地址碰撞概率

- 调整模块加载与测试断言，覆盖共享 helper 和 17 位日期时间格式

Constraint: 回应 PR #280 中关于兼容性、重复逻辑和同秒碰撞风险的评审意见

Confidence: high

Scope-risk: narrow

Tested: node --test tests/cloudmail-provider.test.js tests/background-generated-email-module.test.js

Tested: npm test
