+++ background/cloudmail-provider.js
@@ -1,6 +1,6 @@
 (function cloudMailProviderModule(root, factory) {
-  root.MultiPageBackgroundCloudMailProvider = factory();
-})(typeof self !== 'undefined' ? self : globalThis, function createCloudMailProviderModule() {
+  root.MultiPageBackgroundCloudMailProvider = factory(root);
+})(typeof self !== 'undefined' ? self : globalThis, function createCloudMailProviderModule(root = globalThis) {
   function createCloudMailProvider(deps = {}) {
     const {
       addLog = async () => {},
@@ -18,6 +18,7 @@
       normalizeCloudMailDomains,
       normalizeCloudMailMailApiMessages,
       persistRegistrationEmailState = null,
+      buildRandomNameDateTimeLocalPart = root.MultiPageEmailLocalPartHelpers?.buildRandomNameDateTimeLocalPart,
       pickVerificationMessageWithTimeFallback,
       setEmailState = async () => {},
       setPersistentSettings = async () => {},
@@ -172,46 +173,10 @@
       return chars.join('');
     }
 
-    const ENGLISH_NAME_PREFIXES = [
-      'james', 'john', 'robert', 'michael', 'william', 'david', 'richard', 'joseph',
-      'thomas', 'charles', 'mary', 'patricia', 'jennifer', 'linda', 'elizabeth',
-      'barbara', 'susan', 'jessica', 'sarah', 'karen', 'daniel', 'matthew',
-      'anthony', 'mark', 'donald', 'steven', 'paul', 'andrew', 'joshua', 'kevin',
-      'brian', 'george', 'edward', 'ronald', 'timothy', 'jason', 'jeffrey', 'ryan',
-      'jacob', 'gary', 'nicholas', 'eric', 'jonathan', 'stephen', 'larry', 'justin',
-      'scott', 'brandon', 'benjamin', 'samuel', 'gregory', 'alexander', 'patrick',
-      'frank', 'raymond', 'jack', 'dennis', 'jerry', 'tyler', 'aaron', 'henry',
-      'douglas', 'peter', 'adam', 'zachary', 'nathan', 'walter', 'harold', 'kyle',
-      'carl', 'arthur', 'gerald', 'roger', 'alice', 'emma', 'olivia', 'sophia',
-      'isabella', 'mia', 'amelia', 'harper', 'evelyn', 'abigail', 'emily', 'ella',
-      'scarlett', 'grace', 'chloe', 'victoria', 'riley', 'aria', 'lily', 'nora',
-    ];
-
-    function pickRandomEnglishNamePrefix() {
-      return ENGLISH_NAME_PREFIXES[Math.floor(Math.random() * ENGLISH_NAME_PREFIXES.length)] || 'james';
-    }
-
-    function formatCloudMailDateTimeDigits(date = new Date()) {
-      const current = new Date(date);
-      if (Number.isNaN(current.getTime())) {
-        return '';
-      }
-      const year = String(current.getFullYear());
-      const month = String(current.getMonth() + 1).padStart(2, '0');
-      const day = String(current.getDate()).padStart(2, '0');
-      const hour = String(current.getHours()).padStart(2, '0');
-      const minute = String(current.getMinutes()).padStart(2, '0');
-      const second = String(current.getSeconds()).padStart(2, '0');
-      return `${year}${month}${day}${hour}${minute}${second}`;
-    }
-
     function buildCloudMailNameDateTimeLocalPart(date = new Date()) {
-      const normalizedName = pickRandomEnglishNamePrefix();
-      const dateTimeDigits = formatCloudMailDateTimeDigits(date);
-      if (!dateTimeDigits) {
-        return '';
-      }
-      return `${normalizedName}${dateTimeDigits}`;
+      return typeof buildRandomNameDateTimeLocalPart === 'function'
+        ? buildRandomNameDateTimeLocalPart(date)
+        : '';
     }
 
     async function fetchCloudMailAddress(state, options = {}) {
